portfolio_bot/requirements.txt
ccxt>=4.0.0
python-telegram-bot>=20.0
apscheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
